+++ packages/components/CHANGELOG.md
@@ -1,3 +1,7 @@
+### Enhancements
+
+- Add `BigSkyLogo.Mark` component ([#103612](https://github.com/Automattic/wp-calypso/pull/103612)).
+
 ## 2.3.0
 
 ### Breaking changes